+++ app/Modules/Teams/Resources/Views/index.blade.php
@@ -6,9 +6,9 @@
             <a href="{!! url('teams/'.$team->id.'/'.$team->slug) !!}">
                 <h2>{{ $team->title }}</h2>
 
-                @if ($team->image)
+                @if ($team->banner)
                     <div class="image">
-                        <img class="img-responsive" src="{!! $team->uploadPath().$team->image !!}" alt="{{ $team->title }}">
+                        <img class="img-responsive" src="{!! $team->uploadPath().$team->banner !!}" alt="{{ $team->title }}">
                     </div>
                 @endif
             </a>